pyproject.toml
[build-system]
requires = ["setuptools==70.0.0", "wheel==0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tref"
version = "0.2.0"
description = "Offline-first terminal and Python reference retrieval with versioned KB indexes"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Tref Contributors" }]
license = { text = "MIT" }
dependencies = [
  "typer==0.12.3",
  "rich==13.7.1",
  "fastembed==0.3.6",
  "faiss-cpu==1.8.0.post1",
  "httpx==0.27.2",
  "pyyaml==6.0.2",
  "frontmatter==3.0.8"
]

[project.urls]
Homepage = "https://github.com/tref-org/tref"
Repository = "https://github.com/tref-org/tref"

[project.scripts]
tref = "tref.cli:run"

[tool.setuptools]
packages = ["tref"]

[tool.setuptools.package-data]
tref = ["py.typed"]
